Created TestFunctions static class for getBlobServiceClient and associated methods
