+++ extensions/azure/blobstorage/blob-provision/src/test/java/org/eclipse/dataspaceconnector/provision/azure/blob/ObjectContainerStatusCheckerIntegrationTest.java
@@ -18,6 +18,7 @@ import net.jodah.failsafe.RetryPolicy;
 import org.eclipse.dataspaceconnector.azure.blob.core.AzureBlobStoreSchema;
 import org.eclipse.dataspaceconnector.azure.blob.core.api.BlobStoreApiImpl;
 import org.eclipse.dataspaceconnector.azure.testfixtures.AbstractAzureBlobTest;
+import org.eclipse.dataspaceconnector.azure.testfixtures.TestFunctions;
 import org.eclipse.dataspaceconnector.azure.testfixtures.annotations.AzureStorageIntegrationTest;
 import org.eclipse.dataspaceconnector.common.testfixtures.TestUtils;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
@@ -49,7 +50,7 @@ class ObjectContainerStatusCheckerIntegrationTest extends AbstractAzureBlobTest 
         Vault vault = mock(Vault.class);
 
         when(vault.resolveSecret(account1Name + "-key1")).thenReturn(account1Key);
-        var blobStoreApi = new BlobStoreApiImpl(vault, getEndpoint(account1Name));
+        var blobStoreApi = new BlobStoreApiImpl(vault, TestFunctions.getBlobServiceTestEndpoint(account1Name));
         checker = new ObjectContainerStatusChecker(blobStoreApi, policy);
     }
